Cleanup: Pull out PDF code into its own file
